backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.1
google-generativeai==0.3.1
python-multipart==0.0.6
asyncpg==0.29.0
pydantic[binary]==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4